use int64 for `request_id`
mpv uses int64 for `request_id` internally[1]. Right now, once
`lastRequest` goes past 2**63 - 1, mpv will complain about `request_id`
not being an integer:

> 'request_id' must be an integer. Using other types is deprecated and
> will trigger an error in the future!

and will respond with `request_id` that looks like a float (e.g.
`9223372036854775808.000000`). We fail to parse it as `uint` and thus
match the response to the request, which causes the request to hang.

If we use `int64`, `lastRequest` will wrap after going past 2**63 - 1
and continue to increase from the negative side.

[1]: https://github.com/mpv-player/mpv/blob/2a969a833f98b3486be8007972887c27c08b6f61/input/ipc.c#L124